Fix the comparison for scheduling with the real q
Using just the amount available from the kingdom causes issues as they
may actually already exist in the schedule queue in the state - because
we didn't select them. So we should check the total available queue
length not just what it returned

There is still a bug - we get a different number of moons every time,
and it seems too high. Going to finish putting all the moons in then
will see what the issue is